+++ iti-react/Hooks/Query/index.ts
@@ -1,0 +1,2 @@
+﻿export * from './UsePagination'
+export * from './UseParameterizedQuery'