+++ iti-react/Hooks/index.ts
@@ -1,3 +1,4 @@
 ﻿export * from './UseCancellablePromiseCleanup'
 export * from './UseFieldValidity'
 export * from './UseReadiness'
+export * from './Query'